requirements.txt
backports-abc==0.5
backports.functools-lru-cache==1.6.6
beautifulsoup4==4.9.3
bs4==0.0.2
certifi==2021.10.8
chardet==4.0.0
click==7.1.2
colorama==0.4.6
futures==3.4.0
idna==2.10
Jinja2==2.11.3
livereload==2.6.3
Markdown==3.1.1
MarkupSafe==1.1.1
python-dotenv==0.18.0
python-sourcemaps==1.1
PyYAML==5.4.1
requests==2.27.1
singledispatch==3.7.0
six==1.16.0
soupsieve==1.9.6
tornado==5.1.1
typing==3.10.0.0
urllib3==1.26.19
